telegram_bot_render_ready/requirements.txt
python-telegram-bot==20.3
python-dotenv==1.0.1
aiogram==2.25.1